# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Bank Card Management API
  version: 1.0.0
  description: API для управления банковскими картами с поддержкой аутентификации, авторизации, переводов между картами и административного управления.
  contact:
    name: Bank Card Management
    email: support@bank.com
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html

servers:
  - url: http://localhost:8081
    description: Local development server

tags:
  - name: Authentication
    description: Аутентификация и регистрация пользователей
  - name: Cards
    description: Управление банковскими картами (для пользователей)
  - name: Transfers
    description: Переводы между картами
  - name: Admin Cards
    description: Административное управление картами
  - name: Admin Users
    description: Административное управление пользователями

security:
  - bearerAuth: []

paths:
  /api/auth/signin:
    post:
      tags:
        - Authentication
      summary: Вход в систему
      description: Аутентификация пользователя и получение JWT токена
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Успешная аутентификация
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JwtResponse'
        '401':
          description: Неверные учетные данные
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/auth/signup:
    post:
      tags:
        - Authentication
      summary: Регистрация нового пользователя
      description: Создание нового пользователя в системе
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '200':
          description: Пользователь успешно зарегистрирован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '409':
          description: Конфликт (username или email уже используется)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '400':
          description: Ошибка валидации
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'

  /api/cards:
    post:
      tags:
        - Cards
      summary: Создать новую карту
      description: Создание новой банковской карты для текущего пользователя
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCardRequest'
      responses:
        '200':
          description: Карта успешно создана
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardResponse'
        '400':
          description: Ошибка валидации
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        '409':
          description: Карта с таким номером уже существует
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/cards/my:
    get:
      tags:
        - Cards
      summary: Получить список своих карт
      description: Получение списка карт текущего пользователя с пагинацией и поиском
      parameters:
        - name: page
          in: query
          description: Номер страницы (начиная с 0)
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          description: Размер страницы
          schema:
            type: integer
            default: 10
        - name: sortBy
          in: query
          description: Поле для сортировки
          schema:
            type: string
            default: id
        - name: search
          in: query
          description: Поиск по владельцу или номеру карты
          schema:
            type: string
      responses:
        '200':
          description: Список карт
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardResponsePage'

  /api/cards/my/{id}:
    get:
      tags:
        - Cards
      summary: Получить карту по ID
      description: Получение конкретной карты текущего пользователя
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Карта найдена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardResponse'
        '404':
          description: Карта не найдена или доступ запрещен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      tags:
        - Cards
      summary: Заблокировать свою карту
      description: Запрос на блокировку карты текущего пользователя
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Карта успешно заблокирована
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardResponse'
        '404':
          description: Карта не найдена или доступ запрещен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Невозможно заблокировать истекшую карту
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/transfers:
    post:
      tags:
        - Transfers
      summary: Перевод между своими картами
      description: Перевод средств между картами текущего пользователя
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransferRequest'
      responses:
        '200':
          description: Перевод успешно выполнен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransferResponse'
        '400':
          description: Ошибка валидации
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        '404':
          description: Карта не найдена или доступ запрещен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Ошибка перевода (недостаточно средств, карта неактивна и т.д.)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/admin/cards:
    get:
      tags:
        - Admin Cards
      summary: Получить все карты
      description: Получение списка всех карт в системе (только для администраторов)
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          description: Номер страницы (начиная с 0)
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          description: Размер страницы
          schema:
            type: integer
            default: 10
        - name: sortBy
          in: query
          description: Поле для сортировки
          schema:
            type: string
            default: id
      responses:
        '200':
          description: Список всех карт
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminCardResponsePage'
        '403':
          description: Доступ запрещен (требуется роль ADMIN)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      tags:
        - Admin Cards
      summary: Создать карту (админ)
      description: Создание новой карты администратором
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCardRequest'
      responses:
        '200':
          description: Карта успешно создана
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardResponse'
        '403':
          description: Доступ запрещен (требуется роль ADMIN)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/admin/cards/{id}:
    get:
      tags:
        - Admin Cards
      summary: Получить карту по ID (админ)
      description: Получение конкретной карты с информацией о владельце
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Карта найдена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminCardResponse'
        '404':
          description: Карта не найдена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Доступ запрещен (требуется роль ADMIN)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - Admin Cards
      summary: Удалить карту
      description: Удаление карты из системы
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Карта успешно удалена
        '404':
          description: Карта не найдена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Доступ запрещен (требуется роль ADMIN)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/admin/cards/{id}/activate:
    post:
      tags:
        - Admin Cards
      summary: Активировать карту
      description: Активация заблокированной карты
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Карта успешно активирована
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminCardResponse'
        '404':
          description: Карта не найдена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Невозможно активировать истекшую карту
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Доступ запрещен (требуется роль ADMIN)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/admin/cards/{id}/block:
    post:
      tags:
        - Admin Cards
      summary: Заблокировать карту
      description: Блокировка карты администратором
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Карта успешно заблокирована
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminCardResponse'
        '404':
          description: Карта не найдена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Невозможно заблокировать истекшую карту
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Доступ запрещен (требуется роль ADMIN)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/admin/users:
    get:
      tags:
        - Admin Users
      summary: Получить всех пользователей
      description: Получение списка всех пользователей в системе
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          description: Номер страницы (начиная с 0)
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          description: Размер страницы
          schema:
            type: integer
            default: 10
        - name: sortBy
          in: query
          description: Поле для сортировки
          schema:
            type: string
            default: id
      responses:
        '200':
          description: Список пользователей
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponsePage'
        '403':
          description: Доступ запрещен (требуется роль ADMIN)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/admin/users/{id}:
    get:
      tags:
        - Admin Users
      summary: Получить пользователя по ID
      description: Получение информации о конкретном пользователе
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Пользователь найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '404':
          description: Пользователь не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Доступ запрещен (требуется роль ADMIN)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - Admin Users
      summary: Удалить пользователя
      description: Удаление пользователя из системы
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Пользователь успешно удален
        '404':
          description: Пользователь не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Доступ запрещен (требуется роль ADMIN)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/admin/users/{id}/block:
    post:
      tags:
        - Admin Users
      summary: Заблокировать пользователя
      description: Блокировка пользователя (заблокированный пользователь не сможет войти в систему)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Пользователь успешно заблокирован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '404':
          description: Пользователь не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Доступ запрещен (требуется роль ADMIN)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/admin/users/{id}/unblock:
    post:
      tags:
        - Admin Users
      summary: Разблокировать пользователя
      description: Разблокировка пользователя
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Пользователь успешно разблокирован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '404':
          description: Пользователь не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Доступ запрещен (требуется роль ADMIN)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT токен авторизации. Получите токен через /api/auth/signin

  schemas:
    LoginRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          description: Имя пользователя
          example: user1
        password:
          type: string
          format: password
          description: Пароль
          example: password123

    RegisterRequest:
      type: object
      required:
        - username
        - password
        - email
        - firstName
        - lastName
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 50
          description: Имя пользователя (от 3 до 50 символов)
          example: user1
        password:
          type: string
          format: password
          minLength: 6
          description: Пароль (минимум 6 символов)
          example: password123
        email:
          type: string
          format: email
          description: Email адрес
          example: user1@example.com
        firstName:
          type: string
          description: Имя
          example: John
        lastName:
          type: string
          description: Фамилия
          example: Doe

    JwtResponse:
      type: object
      properties:
        token:
          type: string
          description: JWT токен
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type:
          type: string
          description: Тип токена
          example: Bearer
        id:
          type: integer
          format: int64
          description: ID пользователя
          example: 1
        username:
          type: string
          description: Имя пользователя
          example: user1
        email:
          type: string
          description: Email адрес
          example: user1@example.com
        roles:
          type: array
          items:
            type: string
          description: Роли пользователя
          example: ["USER"]

    UserResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: ID пользователя
          example: 1
        username:
          type: string
          description: Имя пользователя
          example: user1
        email:
          type: string
          description: Email адрес
          example: user1@example.com
        firstName:
          type: string
          description: Имя
          example: John
        lastName:
          type: string
          description: Фамилия
          example: Doe
        roles:
          type: array
          items:
            type: string
          description: Роли пользователя
          example: ["USER"]
        status:
          $ref: '#/components/schemas/UserStatus'

    CreateCardRequest:
      type: object
      required:
        - cardNumber
        - owner
        - expiryDate
        - initialBalance
      properties:
        cardNumber:
          type: string
          description: Номер карты
          example: "1234567890123456"
        owner:
          type: string
          description: Имя владельца карты
          example: "John Doe"
        expiryDate:
          type: string
          format: date
          description: Дата истечения срока действия
          example: "2025-12-31"
        initialBalance:
          type: number
          format: decimal
          minimum: 0
          description: Начальный баланс
          example: 1000.00

    CardResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: ID карты
          example: 1
        cardNumber:
          type: string
          description: Маскированный номер карты
          example: "**** **** **** 3456"
        owner:
          type: string
          description: Имя владельца карты
          example: "John Doe"
        expiryDate:
          type: string
          format: date
          description: Дата истечения срока действия
          example: "2025-12-31"
        status:
          $ref: '#/components/schemas/CardStatus'
        balance:
          type: number
          format: decimal
          description: Баланс карты
          example: 1000.50

    AdminCardResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: ID карты
          example: 1
        cardNumber:
          type: string
          description: Маскированный номер карты
          example: "**** **** **** 3456"
        owner:
          type: string
          description: Имя владельца карты
          example: "John Doe"
        expiryDate:
          type: string
          format: date
          description: Дата истечения срока действия
          example: "2025-12-31"
        status:
          $ref: '#/components/schemas/CardStatus'
        balance:
          type: number
          format: decimal
          description: Баланс карты
          example: 1000.50
        userId:
          type: integer
          format: int64
          description: ID владельца карты
          example: 1

    TransferRequest:
      type: object
      required:
        - fromCardId
        - toCardId
        - amount
      properties:
        fromCardId:
          type: integer
          format: int64
          description: ID карты отправителя
          example: 1
        toCardId:
          type: integer
          format: int64
          description: ID карты получателя
          example: 2
        amount:
          type: number
          format: decimal
          minimum: 0
          exclusiveMinimum: true
          description: Сумма перевода (должна быть положительной)
          example: 100.00

    TransferResponse:
      type: object
      properties:
        transactionId:
          type: string
          format: uuid
          description: ID транзакции
          example: "550e8400-e29b-41d4-a716-446655440000"
        message:
          type: string
          description: Сообщение о результате операции
          example: "Transfer completed successfully"

    CardStatus:
      type: string
      enum:
        - ACTIVE
        - BLOCKED
        - EXPIRED
      description: Статус карты
      example: ACTIVE

    UserStatus:
      type: string
      enum:
        - ACTIVE
        - BLOCKED
      description: Статус пользователя
      example: ACTIVE

    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
          description: Время возникновения ошибки
          example: "2024-01-01T12:00:00"
        status:
          type: integer
          description: HTTP статус код
          example: 404
        error:
          type: string
          description: Тип ошибки
          example: "Not Found"
        message:
          type: string
          description: Сообщение об ошибке
          example: "Card with id 1 not found"
        path:
          type: string
          description: Путь запроса
          example: "/api/cards/my/1"

    ValidationErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
          description: Время возникновения ошибки
          example: "2024-01-01T12:00:00"
        status:
          type: integer
          description: HTTP статус код
          example: 400
        error:
          type: string
          description: Тип ошибки
          example: "Validation Failed"
        message:
          type: object
          additionalProperties:
            type: string
          description: Ошибки валидации по полям
          example:
            cardNumber: "Card number is required"
            amount: "Amount must be positive"
        path:
          type: string
          description: Путь запроса
          example: "/api/cards"

    CardResponsePage:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/CardResponse'
        totalElements:
          type: integer
          description: Общее количество элементов
        totalPages:
          type: integer
          description: Общее количество страниц
        size:
          type: integer
          description: Размер страницы
        number:
          type: integer
          description: Номер текущей страницы
        first:
          type: boolean
          description: Первая страница
        last:
          type: boolean
          description: Последняя страница

    AdminCardResponsePage:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/AdminCardResponse'
        totalElements:
          type: integer
          description: Общее количество элементов
        totalPages:
          type: integer
          description: Общее количество страниц
        size:
          type: integer
          description: Размер страницы
        number:
          type: integer
          description: Номер текущей страницы
        first:
          type: boolean
          description: Первая страница
        last:
          type: boolean
          description: Последняя страница

    UserResponsePage:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/UserResponse'
        totalElements:
          type: integer
          description: Общее количество элементов
        totalPages:
          type: integer
          description: Общее количество страниц
        size:
          type: integer
          description: Размер страницы
        number:
          type: integer
          description: Номер текущей страницы
        first:
          type: boolean
          description: Первая страница
        last:
          type: boolean
          description: Последняя страница